refactor(chat): dedupe frontend store + gateway ergo helpers, codec polish
Cleanup pass over the chat surface.

astro-irc:
- extract the ~90% identical state.ts (embed) and service.ts (main app)
  into a shared src/lib/chat/store.ts (types, atoms, computed stores,
  pushMessage/makeId/systemMessage/switchChannel, and the IRC handleIncoming
  reducer). The diverging bits — PONG transport + 001 auto-join — are injected
  via an IncomingHandlers callback. Both modules are now thin wrappers.
- collapse the two byte-identical format.ts copies into src/lib/chat/format.ts.
- drop the unused $canSend atom from the main-app service module.

irc-gateway:
- extract the duplicated parse_privmsg (minechat + history) and the ergo
  host/port/ws-url env lookups (repeated across minechat, history, rest,
  websocket, irc) into a shared gateway/ergo.rs with connect_irc()/parse_privmsg.
  Behaviour-identical: same env vars, same defaults.

@kbve/chat:
- tighten GamechatFrame.payload to Record<string, unknown>.
- add parseEnvelope edge-case tests (@-prefixed sender, nested-object payload,
  empty platform, non-object JSON tail).
- README: document parseIrcLine/decodeIrcTrailing + the gamechat subpath +
  full platform wire-name table; fix the stale @kbve/proto note (now vendored).

Verified: chat lint + 47 tests, astro-irc embed + full astro build + tsc clean,
irc-gateway 48 tests, laser nx build.
