+++ apps/irc/astro-irc/src/lib/chat/format.ts
