+++ apps/irc/irc-gateway/src/gateway/mod.rs
@@ -1,3 +1,4 @@
+pub mod ergo;
 pub mod history;
 pub mod irc;
 pub mod kv;
